backend/requirements.txt
fastapi
uvicorn
pdfplumber
pandas
python-multipart
pytest-cov